+++ src/main/java/com/flytrap/rssreader/presentation/controller/FolderUpdateController.java
@@ -10,9 +10,9 @@ import com.flytrap.rssreader.presentation.dto.SessionMember;
 import com.flytrap.rssreader.presentation.dto.SubscribeRequest;
 import com.flytrap.rssreader.presentation.resolver.Login;
 import com.flytrap.rssreader.service.alert.AlertService;
-import com.flytrap.rssreader.service.FolderSubscribeService;
-import com.flytrap.rssreader.service.FolderUpdateService;
-import com.flytrap.rssreader.service.FolderVerifyOwnerService;
+import com.flytrap.rssreader.service.folder.FolderSubscribeService;
+import com.flytrap.rssreader.service.folder.FolderUpdateService;
+import com.flytrap.rssreader.service.folder.FolderVerifyOwnerService;
 import com.flytrap.rssreader.service.SubscribeService;
 import jakarta.validation.Valid;
 import java.util.List;
