+++ src/main/java/com/flytrap/rssreader/presentation/controller/PostUpdateController.java
@@ -4,7 +4,8 @@ import com.flytrap.rssreader.global.model.ApplicationResponse;
 import com.flytrap.rssreader.presentation.dto.ReactionRequest;
 import com.flytrap.rssreader.presentation.dto.SessionMember;
 import com.flytrap.rssreader.presentation.resolver.Login;
-import com.flytrap.rssreader.service.FolderVerifyOwnerService;
+import com.flytrap.rssreader.service.SharedFolderReadService;
+import com.flytrap.rssreader.service.folder.FolderVerifyOwnerService;
 import com.flytrap.rssreader.service.PostOpenService;
 import com.flytrap.rssreader.service.ReactionService;
 import jakarta.validation.Valid;
