+++ src/main/java/com/flytrap/rssreader/presentation/controller/SharedFolderUpdateController.java
@@ -6,8 +6,8 @@ import com.flytrap.rssreader.global.model.ApplicationResponse;
 import com.flytrap.rssreader.presentation.dto.InviteMemberRequest;
 import com.flytrap.rssreader.presentation.dto.SessionMember;
 import com.flytrap.rssreader.presentation.resolver.Login;
-import com.flytrap.rssreader.service.FolderUpdateService;
-import com.flytrap.rssreader.service.FolderVerifyOwnerService;
+import com.flytrap.rssreader.service.folder.FolderUpdateService;
+import com.flytrap.rssreader.service.folder.FolderVerifyOwnerService;
 import com.flytrap.rssreader.service.MemberService;
 import com.flytrap.rssreader.service.SharedFolderUpdateService;
 import javax.security.sasl.AuthenticationException;
